feat: 쿠폰 종류 컴포넌트 분리 & 중지 쿠폰 레이아웃
