+++ src/components/CouponList/CouponItem/CouponExpired/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const CouponExpired = () => {
+  return <div></div>;
+};
+
+export default CouponExpired;
