+++ src/components/CouponList/CouponItem/CouponWait/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const CouponWait = () => {
+  return <div></div>;
+};
+
+export default CouponWait;
